Changed: Refactor timeout validation to support user-configurable limits
- Standardize all timeout types from u64 to u32 for memory efficiency
- Move timeout validation into core library for both webfetch and bash tools
- Add max_timeout_ms to BashTool and WebFetchTool for LLM constraint enforcement
- Update tool APIs: with_settings() for webfetch, with_default_timeout_ms()/with_max_timeout_ms() for bash
- Remove redundant validation from serdesai layer, core now handles all timeout validation
- Update execute_command signatures to accept u32 timeout_ms and max_timeout_ms parameters
- Update agent settings and all tests for new API

Allows users to configure both default and maximum timeout limits,
restricting LLM from exceeding max while giving users full control.
Bash tool now matches webfetch architecture where core handles all validation.

+++ src/llm-coding-tools-agents/src/types/tool_settings.rs
@@ -143,7 +143,7 @@ pub struct BashToolSettings {
         default = "bash_default_timeout_ms",
         deserialize_with = "deserialize_min_timeout_ms"
     )]
-    pub timeout_ms: u64,
+    pub timeout_ms: u32,
 }
 
 impl Default for BashToolSettings {
@@ -163,7 +163,7 @@ pub struct WebFetchToolSettings {
         default = "webfetch_default_timeout_ms",
         deserialize_with = "deserialize_min_timeout_ms"
     )]
-    pub timeout_ms: u64,
+    pub timeout_ms: u32,
     /// Maximum response size in MiB (default: 5, min: 1).
     #[serde(
         default = "webfetch_default_max_response_size_mib",
@@ -213,12 +213,12 @@ const fn glob_default_limit() -> usize {
 }
 
 #[inline]
-const fn bash_default_timeout_ms() -> u64 {
+const fn bash_default_timeout_ms() -> u32 {
     bash::DEFAULT_TIMEOUT_MS
 }
 
 #[inline]
-const fn webfetch_default_timeout_ms() -> u64 {
+const fn webfetch_default_timeout_ms() -> u32 {
     webfetch::DEFAULT_TIMEOUT_MS
 }
 
@@ -272,11 +272,11 @@ where
     Ok(value)
 }
 
-fn deserialize_min_timeout_ms<'de, D>(deserializer: D) -> Result<u64, D::Error>
+fn deserialize_min_timeout_ms<'de, D>(deserializer: D) -> Result<u32, D::Error>
 where
     D: serde::Deserializer<'de>,
 {
-    let value = u64::deserialize(deserializer)?;
+    let value = u32::deserialize(deserializer)?;
     if value < MIN_TIMEOUT_MS {
         return Err(serde::de::Error::custom(format!(
             "value must be >= {}",
