+++ src/llm-coding-tools-core/src/tool_metadata/bash.rs
@@ -6,10 +6,10 @@ use super::ParamMetadata;
 pub const NAME: &str = "bash";
 
 /// Default timeout in milliseconds.
-pub const DEFAULT_TIMEOUT_MS: u64 = 120_000;
+pub const DEFAULT_TIMEOUT_MS: u32 = 120_000;
 
 /// Maximum timeout in milliseconds.
-pub const MAX_TIMEOUT_MS: u64 = 600_000;
+pub const MAX_TIMEOUT_MS: u32 = 600_000;
 
 /// Tool description.
 pub const DESCRIPTION: &str = "Run a shell command in a fresh process.";
