+++ src/llm-coding-tools-core/src/util.rs
@@ -16,7 +16,7 @@ pub const LIKELY_CHARS_PER_LINE_MAX: usize = ESTIMATED_CHARS_PER_LINE * 4;
 pub const MIN_LIMIT: usize = 1;
 
 /// Minimum value for timeout fields in milliseconds (e.g., bash.timeout_ms, webfetch.timeout_ms).
-pub const MIN_TIMEOUT_MS: u64 = 1000;
+pub const MIN_TIMEOUT_MS: u32 = 1000;
 
 /// Formats a line with its line number for output.
 ///
